JSON WEB TOKEN & Middlewares
Today we have applied jwt (JSON Web Token) and applied it in the form of middleware so that we can verify the user before calling each and every API so that it provides security from unauthorized user. It helps to reduce writing of same code in every API by using middleware.