Super Admin work and Multiple Products through Excel in Backend
Today, we have done work of super admin in which we have built an API in backend to fetch all users except super admin and displayed them on super admin panel and given two action to enable and disable them also. And we have also applied pagination on that page. After that We have read the data from excel file yesterday but we had not built an API for it. Today, along with this work we have also built an API to insert multiple products in database by checking the data type and data structure of the data which are going to insert whether it is declared in schema or not by applying many function in it.