Some changes in backend and Created a shopkeeper through super admin.
I have just done the work to create and verify the account of shopkeeper using super admin panel by verifying through OTP on the mail. For that, I have created an API in backend to get all the shopkeeper for creating executive account and done some changes in the previous API's we have done. Also done work on the Super admin header and changed the structure of the modal on the Super admin page to enhance user experience by creating user account.