Some Important changes both in backend and Frontend
We have fetched Indian States and their cities JSON data in backend and sent it back to frontend to enhance user experience and the UI using an API on the step of creating account of users on the Super Admin Panel. And after that some few changes are also done like giving feature to user to download an excel file format so that it upload it easily and many more option. I have also applied pagination on the All Products page.